changed the scope to be private for the test functions, then relocated the tests as those are now private, which cleans a bit of the files for symmetric/complex and added a note
