+++ spindalis/src/reduction/matrix/francis/complex.rs
@@ -167,100 +167,3 @@ pub fn francis_iteration_cpx_2x2(
     apply_gt_right(hess_lin_matrix, 0, 1, stride, size, cosine, sine);
     apply_g_left(hess_lin_matrix, 0, 1, stride, 2, cosine, sine);
 }
-
-#[cfg(test)]
-mod test_verify_correspondance_complex {
-    use super::*;
-    use crate::reduction::matrix::francis::constants::{MAX_ITERS, TOLERANCE};
-    use crate::reduction::matrix::francis::primitives::hessenberg_lq;
-    use crate::reduction::matrix::francis::verify::{
-        decomp_cpx_with_rotation, hessenberg_lq_with_rotation,
-    };
-    use rand::prelude::*;
-    use rand_distr::StandardNormal;
-
-    fn generate_random_vector(n: usize) -> Vec<f64> {
-        let mut rng = StdRng::seed_from_u64(42);
-        let mut data = vec![0f64; n];
-        for d in data.iter_mut().take(n) {
-            *d = rng.sample(StandardNormal);
-        }
-        data
-    }
-    fn generate_identity_vector(m: usize, n: usize) -> Vec<f64> {
-        let mut vector = vec![0f64; m * n];
-        let mut idx = 0;
-        for _ in 0..m {
-            vector[idx] = 1f64;
-            idx += 1 + n;
-        }
-        vector
-    }
-    // A * A^T so it's complex (with float noise, like the other tests)
-    fn generate_approx_complex_vector(n: usize) -> Vec<f64> {
-        let a = generate_random_vector(n * n);
-        let mut result = vec![0f64; n * n];
-        for i in 0..n {
-            for j in 0..n {
-                let mut sum = 0f64;
-                for k in 0..n {
-                    sum += a[i * n + k] * a[j * n + k];
-                }
-                result[i * n + j] = sum;
-            }
-        }
-        result
-    }
-    fn approx_slice_eq(a: &[f64], b: &[f64], tolerance: f64) -> bool {
-        assert_eq!(a.len(), b.len());
-        a.iter().zip(b.iter()).all(|(x, y)| {
-            if x.is_nan() || y.is_nan() {
-                return false;
-            }
-            (x - y).abs() < tolerance
-        })
-    }
-    #[test]
-    fn test_decomp_cpx_matches_decomp_cpx_with_rotation() {
-        // 5 random shapes
-        for dim in [2, 3, 5, 6, 8] {
-            let (rows, cols) = (dim, dim);
-            let stride = dim;
-
-            let base = generate_approx_complex_vector(dim);
-
-            // --- plain decomp_cpx path ---
-            let mut h_plain = base.clone();
-            let mut p_plain = vec![0f64; cols];
-            let mut w_plain = vec![0f64; rows.max(3)];
-            hessenberg_lq(&mut h_plain, &mut p_plain, &mut w_plain, rows, cols, stride);
-            decomp_cpx(
-                &mut h_plain,
-                &mut p_plain,
-                &mut w_plain,
-                dim,
-                dim,
-                stride,
-                MAX_ITERS,
-                TOLERANCE,
-            );
-
-            // --- rotation-tracking decomp_cpx_with_rotation path ---
-            let mut h_rot = base.clone();
-            let mut r_rot = generate_identity_vector(rows, cols);
-            let mut p_rot = vec![0f64; cols];
-            let mut w_rot = vec![0f64; rows.max(3)];
-            hessenberg_lq_with_rotation(
-                &mut h_rot, &mut p_rot, &mut r_rot, &mut w_rot, rows, cols, stride,
-            );
-            decomp_cpx_with_rotation(
-                &mut h_rot, &mut p_rot, &mut r_rot, &mut w_rot, rows, cols, stride,
-            );
-
-            assert!(
-                approx_slice_eq(&h_plain, &h_rot, 1e-6),
-                "dim={dim}: decomp_cpx and decomp_cpx_with_rotation diverged\nplain: {h_plain:?}\nrot:   {h_rot:?}"
-            );
-        }
-    }
-}
