+++ spindalis/src/reduction/matrix/francis/mod.rs
@@ -4,4 +4,4 @@ pub mod givens;
 pub mod interface;
 pub mod primitives;
 pub mod symmetric;
-pub mod verify; // for testing and reconstruction don't expose
+mod verify;
